fixed auth

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,16 @@ use App\Post;
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -41,11 +51,24 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload
+        if($request->hasFile('cover_image')){
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $fileName = pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $fileName. '_' .time(). '.' .$extension;
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
         
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created');
@@ -72,6 +95,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check the correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'You are not authorized');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -107,6 +134,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+         // Check the correct user
+         if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'You are not authorized');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
